requirements.txt
streamlit
pandas
scikit-learn
numpy<2.0
scikit-surprise